// packages/short-ai-app/src/features/api/generated/models/taskType.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * ShortAI
 * OpenAPI spec version: 0.1.0
 */

/**
 * An enumeration.
 */
export type TaskType = (typeof TaskType)[keyof typeof TaskType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskType = {
  file: 'file',
  text: 'text',
  url: 'url'
} as const